# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: NetPulse API
  version: 0.12.0
  description: Versioned JSON contract for diagnoses, services, incidents, map aggregates, the user platform, the developer workspace, organizations, and the internal admin platform. No invented health. Auth tokens are session ids or hashed API keys, never localStorage.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/diagnoses:
    post:
      summary: Create a diagnosis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target]
              properties:
                target:
                  type: string
      responses:
        "201":
          description: Queued
        "400":
          description: Validation error
        "403":
          description: SSRF blocked
        "429":
          description: Rate limited
  /v1/diagnoses/{id}:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Diagnosis
        "404":
          description: Not found
...
  /v1/services/{slug}:
    get:
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Catalog entry
        "404":
          description: Not found
  /v1/incidents:
    get:
      summary: Stored incidents only
      parameters:
        - in: query
          name: service
          schema: { type: string }
        - in: query
          name: region
          schema: { type: string }
        - in: query
          name: network
          schema: { type: string }
        - in: query
          name: severity
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: time
          schema: { type: string, enum: [all, 24h, 7d, 30d] }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: sort
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
      responses:
        "200":
          description: Stored incidents only. Empty means none recorded.
  /v1/incidents/{id}:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stored incident
        "404":
          description: Not found
  /v1/map/aggregates:
    get:
      summary: Coarse geographic aggregates
      description: >
        Returns clustered, viewport-capped cells derived from stored incidents.
        Never includes raw measurements or precise individual locations.
        Coordinates, when present, are snapped to 1 degree.
      parameters:
        - in: query
          name: level
          schema: { type: string, enum: [world, country, region, network, service] }
        - in: query
          name: parent
          schema: { type: string }
        - in: query
          name: west
          schema: { type: number }
        - in: query
          name: south
          schema: { type: number }
        - in: query
          name: east
          schema: { type: number }
        - in: query
          name: north
          schema: { type: number }
        - in: query
          name: layers
          schema: { type: string }
          description: Comma-separated global,regional,network,service,incidents
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: service
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, maximum: 250 }
      responses:
        "200":
          description: Aggregated cells only. Empty means none stored.
        "503":
          description: Store unavailable
  /v1/auth/register:
    post:
      summary: Create an account
      responses:
        "201":
          description: Account created. sessionToken is for the BFF cookie only.
        "409":
          description: Email already registered
  /v1/auth/login:
    post:
      summary: Sign in
      responses:
        "200":
          description: Session established
        "401":
          description: Generic email or password failure
...
  /v1/me/diagnoses:
    get:
      summary: Current user's diagnoses only
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: target
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string }
        - in: query
          name: to
          schema: { type: string }
...
  /v1/admin/me:
    get:
      summary: Operator identity. Signed-in non-operators receive 404.
      responses:
        "200":
          description: Operator
        "401":
          description: Unauthenticated
        "404":
          description: Not found
...